balance: stretch every Discernment sink across the real endgame
Discernment was a dead stat for the whole late game. Every sink it fed
saturated inside the first half-million while endgame Discernment runs to
15M-100M, so past ~5k a point in Attack was strictly better than a point in
Discernment and the stat stopped being a choice.

Crit saturation moves out 100x. The CAPS are untouched (+10pp chance,
+1.0x damage); only the per-point rate changed, so crit chance now tops out
at 100,000 Discernment instead of 1,000, and crit damage at 500,000 instead
of 5,000. This costs the early game: 1,000 Discernment is now worth +0.1pp,
not a capped +10pp.

Rare drops re-anchor on 1M and stop at 10M. Halfway 50,000 -> 111,111 puts
90% of the bonus at 1,000,000, and a new RareDropDiscernmentCap clamps the
input at 10,000,000 so the chase has a stated end instead of an asymptote
nobody reaches.

Clear ladders run to 5,000, per raid and per zone. The zone ladder grew
from six rungs to nine. Raids had no per-raid ladder at all -- only one
GLOBAL RaidCompletions counter with two authored rungs (10 and 100), so a
whole raid roster shared a ladder that finished in an afternoon. RaidClears
is a new scoped metric that expands across the raid roster at boot exactly
as the zone ladder expands across quest zones, routed by raid definition id,
with the active-raid id and player in the referenceId so a re-processed kill
cannot double-advance it. RaidCompletions stays as the lifetime tally.

Ladder ids are now keyed on threshold rather than rarity. ItemRarity stops
at Orange permanently, so a nine-rung ladder necessarily repeats a rarity
and rarity-keyed ids would collide and fail the boot validator. NOTE: this
renames every existing zone-rerun achievement id, so in-flight progress rows
do not carry over -- harmless ahead of the planned account wipe.

GetCritProfile now takes a long. The (int) cast that fed it was the thinnest
numeric margin in the game -- 21x headroom at 100M Discernment against an
int32 ceiling, where a wrap would have silently clamped crit to base rather
than failing loudly.

1143 unit tests pass, 0 warnings.

+++ src/ROTA.Application/Interfaces/IStatService.cs
@@ -19,5 +19,5 @@ public interface IStatService
 
     int XpToNextLevel(int level);
 
-    CritProfile GetCritProfile(int discernment);
+    CritProfile GetCritProfile(long discernment);
 }
